# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tangible",
    "description": "Self-hosted disc-image preservation, management, and burning.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/burn-attempts/{attempt_id}/complete": {
      "post": {
        "tags": [
          "workers"
        ],
        "summary": "Record the end of an attempt.",
        "description": "Record the end of an attempt. Idempotent: retrying after a lost response returns the same physical copy rather than recording a second disc.",
        "operationId": "complete",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "description": "Attempt identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such leased attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "worker_credential": []
          }
        ]
      }
    },
...
    "/api/v1/session": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "The signed-in session.",
        "operationId": "current_session",
        "responses": {
          "200": {
            "description": "The session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Sign in.",
        "description": "Sign in with a username and password. Sets the session cookie and returns the anti-forgery token mutations must send.",
        "operationId": "sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "description": "The username or password is wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many failures for this username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Sign out.",
        "description": "# Errors\n\n`STORAGE_UNAVAILABLE` if the session cannot be ended.",
        "operationId": "sign_out",
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/setup": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Whether the server still needs its first account.",
        "description": "# Errors\n\n`STORAGE_UNAVAILABLE` if accounts cannot be counted.",
        "operationId": "setup_status",
        "responses": {
          "200": {
            "description": "Whether setup is needed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create the first administrator and sign them in.",
        "description": "Create the first account, an administrator, and sign it in. Works only while no account exists.",
        "operationId": "complete_setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created and signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "409": {
            "description": "Setup is already complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The username or password will not do",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Every account.",
        "description": "# Errors\n\n`STORAGE_UNAVAILABLE` if accounts cannot be read.",
        "operationId": "list_accounts",
        "responses": {
          "200": {
            "description": "Every account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create an account.",
        "description": "# Errors\n\n`VALIDATION_FAILED` for a username, password or role that will not do;\n`CONFLICT` for a username already in use.",
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The username is taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The username, password or role will not do",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompletionRequest": {
        "type": "object",
        "description": "What a worker sends when it finishes.",
        "required": [
          "lease_token",
          "last_sequence",
          "write_report",
          "physical_medium"
        ],
        "properties": {
          "failure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FailureReport",
                "description": "Why it failed, when it did."
              }
            ]
          },
          "last_sequence": {
            "type": "integer",
            "format": "int64",
            "description": "The last event sequence the worker emitted."
          },
          "lease_token": {
            "type": "string",
            "description": "The lease token from the claim."
          },
          "physical_medium": {
            "$ref": "#/components/schemas/PhysicalMedium",
            "description": "What was in the drive."
          },
          "verification_report": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VerificationReport",
                "description": "The read-back comparison, when one ran."
              }
            ]
          },
          "write_report": {
            "$ref": "#/components/schemas/WriteReport",
            "description": "The engine's account of the write."
          }
        }
      },
      "CompletionResponse": {
        "type": "object",
        "description": "What the server acknowledges.",
        "required": [
          "acknowledged",
          "eject"
        ],
        "properties": {
          "acknowledged": {
            "type": "boolean",
            "description": "Always true when the server has the completion recorded."
          },
          "eject": {
            "type": "boolean",
            "description": "Whether the drive should eject."
          },
          "physical_copy_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The disc recorded, when the attempt consumed media."
          }
        }
      },
...
      "CreateUserRequest": {
        "type": "object",
        "description": "An account to create.",
        "required": [
          "username",
          "password",
          "role"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "At least 12 characters."
          },
          "role": {
            "type": "string",
            "description": "`viewer`, `operator` or `administrator`."
          },
          "username": {
            "type": "string",
            "description": "1 to 64 lowercase letters, digits, `.`, `_` or `-`."
          }
        },
        "additionalProperties": false
      },
      "Credentials": {
        "type": "object",
        "description": "A username and password.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "The password."
          },
          "username": {
            "type": "string",
            "description": "The username. Case does not matter."
          }
        },
        "additionalProperties": false
      },
...
      "FailureReport": {
        "type": "object",
        "description": "Why an attempt did not succeed.\n\nSent alongside the reports rather than inferred from them, because the\nreports say *what* happened and this says why. Without it a failed burn\ntells an operator only that it failed.",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, such as `PREFLIGHT_MEDIUM_NOT_BLANK`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail, bounded by the server before storage."
          }
        }
      },
...
      "PhysicalMedium": {
        "type": "object",
        "description": "What was in the drive.",
        "required": [
          "profile"
        ],
        "properties": {
          "manufacturer_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Manufacturer identifier, when the drive reported one."
          },
          "profile": {
            "type": "string",
            "description": "Media profile, such as `bd-r-25`."
          },
          "serial": {
            "type": [
              "string",
              "null"
            ],
            "description": "Media serial, when the drive reported one."
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "An RFC 9457 problem document.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code. Clients branch on this, not on `detail`."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation. May be reworded between releases."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path, when known."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated in the body as the RFC allows.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type."
          }
        }
      },
...
      "SessionView": {
        "type": "object",
        "description": "The signed-in session.",
        "required": [
          "username",
          "role",
          "csrf_token",
          "expires_at"
        ],
        "properties": {
          "csrf_token": {
            "type": "string",
            "description": "The token every mutation must send in `X-CSRF-Token`.\n\nNot a credential on its own: it is useless without the session\ncookie, which a page's script cannot read."
          },
          "expires_at": {
            "type": "string",
            "description": "When the session ends however much it is used."
          },
          "role": {
            "type": "string",
            "description": "What they may do."
          },
          "username": {
            "type": "string",
            "description": "Who is signed in."
          }
        }
      },
      "SetupStatus": {
        "type": "object",
        "description": "Whether the server still needs its first account.",
        "required": [
          "needed"
        ],
        "properties": {
          "needed": {
            "type": "boolean",
            "description": "True until the first administrator exists."
          }
        }
      },
...
      "UserList": {
        "type": "object",
        "description": "A list of accounts.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserView"
            },
            "description": "Every account, oldest first."
          }
        }
      },
      "UserView": {
        "type": "object",
        "description": "An account, without anything that would sign anybody in.",
        "required": [
          "id",
          "username",
          "role",
          "disabled",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "description": "When it was created."
          },
          "disabled": {
            "type": "boolean",
            "description": "Whether the account has been disabled."
          },
          "id": {
            "type": "string",
            "description": "Opaque identifier."
          },
          "last_signed_in_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "When it last signed in."
          },
          "role": {
            "type": "string",
            "description": "`viewer`, `operator` or `administrator`."
          },
          "username": {
            "type": "string",
            "description": "The username, lowercase."
          }
        }
      },
...
      "WriteReport": {
        "type": "object",
        "description": "The engine's account of the write.",
        "required": [
          "state",
          "engine",
          "engine_version",
          "started_at",
          "completed_at"
        ],
        "properties": {
          "completed_at": {
            "type": "string",
            "description": "When it ended."
          },
          "engine": {
            "type": "string",
            "description": "Which engine wrote."
          },
          "engine_version": {
            "type": "string",
            "description": "That engine's version."
          },
          "started_at": {
            "type": "string",
            "description": "When writing began."
          },
          "state": {
            "type": "string",
            "description": "`success` or a failure description."
          }
        }
      }
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tangible_session",
        "description": "Set by signing in. Mutations must also send the session's anti-forgery token in X-CSRF-Token."
      },
      "worker_credential": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}